🚑 last time seeing rust

+++ src/bot/cat.rs
@@ -1,4 +1,5 @@
 use super::core::MEDIA_GROUP_RESOURCE;
+use super::core::CORE_CONFIG;
 use teloxide::prelude::*;
 
 pub async fn post_cat(b: Bot, msg: Message) -> ResponseResult<Message> {
@@ -26,9 +27,11 @@ pub async fn post_cat(b: Bot, msg: Message) -> ResponseResult<Message> {
                     let target_id = sha256::digest(file_id);
 
                     super::net::upload_oss(file_id, &target_id).await?;
-                    super::net::delete_file(file_id).await?;
+                    if !CORE_CONFIG.lock().await.keep_image_cache {
+                        super::net::delete_file(file_id).await?;
+                    }
 
-                    super::metadata::insert_metadata(caption, vec![target_id]).await?;
+                    super::metadata::insert_metadata(caption, target_id).await?;
                     super::metadata::update_metadata().await?;
 
                     return b
@@ -44,19 +47,20 @@ pub async fn post_cat(b: Bot, msg: Message) -> ResponseResult<Message> {
                     map.remove(&photo.media_group_id.as_ref().unwrap().to_string());
                     drop(map);
 
-                    let mut ids: Vec<String> = Vec::new();
-
                     for file_id in &photos {
                         super::net::download_file(&b, file_id).await?;
 
                         // Use hash as id
                         let target_id = sha256::digest(file_id);
                         super::net::upload_oss(file_id, &target_id).await?;
+                        if !CORE_CONFIG.lock().await.keep_image_cache {
+                            super::net::delete_file(file_id).await?;
+                        }
+
                         super::net::delete_file(file_id).await?;
-                        ids.push(target_id);
+                        super::metadata::insert_metadata(caption, target_id).await?;
                     }
 
-                    super::metadata::insert_metadata(caption, ids).await?;
                     super::metadata::update_metadata().await?;
 
                     return b
